LayoutTests/imported/w3c/web-platform-tests/css/css-page/parsing/margin-rules-002-expected.txt


FAIL margin rule without any preceding @page properties assert_equals: expected 1 but got 0
FAIL margin rule followed by @page properties assert_equals: expected 1 but got 0
FAIL margin rule preceded by @page properties assert_equals: expected 1 but got 0
FAIL margin rules with @page properties between assert_equals: expected 2 but got 0
